third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-garbage-collection-ran-during-error.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let promise;
  return setBluetoothFakeAdapter('MissingServiceHeartRateAdapter')
      .then(
          () =>
              requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]}))
      .then(device => device.gatt.connect())
      .then(gattServer => {
        promise = assert_promise_rejects_with_message(
            gattServer.getPrimaryService('heart_rate'),
            new DOMException(
                'GATT Server is disconnected. ' +
                    'Cannot retrieve services. ' +
                    '(Re)connect first with `device.gatt.connect`.',
                'NetworkError'));
        // Disconnect called to clear attributeInstanceMap and allow the
        // object to get garbage collected.
        gattServer.disconnect();
      })
      .then(runGarbageCollection)
      .then(() => promise);
}, 'Garbage Collection ran during a getPrimaryService call that failed. ' +
   'Should not crash.');

</script>
